+++ test/production/app-dir/next-types-plugin/sync-params-type-check/app/page.tsx
